restructured for reusability
