+++ src/codeoceanauxtools/get_groups.py
@@ -1,26 +1,10 @@
 #!/usr/bin/env python
-import subprocess
 import sys
 
-
-def get_groups(filename="sample_sheet.csv"):
-    filepath = (
-        subprocess.check_output(["find", "../data", "-name", filename])
-        .decode("utf-8")
-        .strip()
-    )
-    groups_set = set()
-    with open(f"{filepath}") as infile:
-        lines = infile.readlines()
-        for line in lines:
-            line = line.strip()
-            line_group = line.split(",")[0]
-            groups_set.add(line_group)
-        return ",".join(sorted(list(groups_set)))
-
+import utils.util
 
 if __name__ == "__main__":
     if len(sys.argv) > 1:
-        print(get_groups(sys.argv[1]))
+        print(utils.util.get_groups(sys.argv[1]))
     else:
-        print(get_groups())
+        print(utils.util.get_groups())
