+++ src/codeoceanauxtools/get_pipeline_confirm.py
@@ -1,10 +1,5 @@
 #!/usr/bin/env python
-import os
-
-
-def is_pipeline():
-    return 1 if bool(os.getenv("AWS_BATCH_JOB_ID")) else 0
-
+import utils.util
 
 if __name__ == "__main__":
-    print(is_pipeline())
+    print(utils.util.is_pipeline())
