+++ src/codeoceanauxtools/get_read_direction.py
@@ -1,13 +1,7 @@
 #!/usr/bin/env python
 import sys
 
-
-def get_read_direction(filepath):
-    filename = filepath.split("/")[-1]
-    if "_" not in filename:
-        return 0
-    return "1" if "1" in filename.split("_")[-1].split(".")[0] else "2"
-
+import utils.util
 
 if __name__ == "__main__":
-    print(get_read_direction(sys.argv[1]))
+    print(utils.util.get_read_direction(sys.argv[1]))
