+++ tests/test_get_read_direction.py
@@ -1,4 +1,4 @@
-from src.codeoceanauxtools.get_read_direction import get_read_direction
+from src.codeoceanauxtools.utils.util import get_read_direction
 
 
 # Tests for get_read_direction
