+++ tests/test_get_rev_file.py
@@ -1,4 +1,4 @@
-from src.codeoceanauxtools.get_rev_file import get_rev_file
+from src.codeoceanauxtools.utils.util import get_rev_file
 
 
 # Tests for get_rev_file
@@ -40,3 +40,8 @@ def test_get_rev_file_R1():
 def test_get_rev_file_R2():
     filepath = "gsm123_rep1_R2.fastq.gz"
     assert get_rev_file(filepath) == "gsm123_rep1_R2.fastq.gz"
+
+
+# def test_get_rev_file_single_end_path():
+#     filepath = "/path/to/reads/gsm123rep1.fastq.gz"
+#     assert get_rev_file(filepath) == "/path/to/reads/gsm123rep1.fastq.gz"
